+++ projects/egeo/src/lib/st-tag/st-tag.model.ts
@@ -14,5 +14,14 @@ export class StTagItem {
    text: string;
    icon?: string;
    bubble?: string;
+   rightIcon?: string;
    [key: string]: any;
+}
+
+export class STagClasses {
+   removable?: boolean;
+   info?: boolean;
+   clickable?: boolean;
+   selectable?: boolean;
+   [key: string]: boolean;
 }